refactor: 기능구현목록 수정사항에 맞게 파일 분리
